requirements.txt
fastapi==0.95.1
uvicorn==0.21.1
# Remove or comment out the specific torch/torchvision lines for now
#torch==1.13.1
#torchvision==0.14.1
Pillow==9.5.0
requests==2.28.2
# Add SQLAlchemy if it wasn't in the original file:
sqlalchemy